+++ AnastasiaMaremshaova_lesson_1/Project/src/index.html
@@ -3,9 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>NikutevDE</title>
+    <title>React</title>
 </head>
 <body>
-<div id="root"></div>
+    <div id="root"></div> 
 </body>
 </html>
