+++ AnastasiaMaremshaova_lesson_1/Project/webpack.config.js
@@ -1,30 +1,30 @@
-const path = require("path");
+const path = require("path"); 
 const HtmlWebpackPlugin = require("html-webpack-plugin");
-
-
 module.exports = {
     entry: path.resolve(__dirname, "src", "index.js"),
     output: {
         path: path.resolve(__dirname, "dist"),
-        filename: "index.js",
+        filename: "index.js"
     },
+
     module: {
+
         rules: [
             {
-                test: /\.(js|jsx)$/,
+                test: /\.(js)$/,
                 include: path.resolve(__dirname, "src"),
                 loader: 'babel-loader',
                 options: {
-                    presets: ['@babel/env', '@babel/react'],
-                    plugins: ['@babel/plugin-proposal-class-properties']
+                    presets: ['@babel/env', '@babel/react']
                 }
-            },
+
+            }
         ]
+
     },
+
     plugins: [
-        new HtmlWebpackPlugin({ template: path.resolve(__dirname, "src", "index.html")})
-    ],
-    resolve: {
-        extensions: [".jsx", ".js"],
-    }
+        new HtmlWebpackPlugin({template: path.resolve(__dirname, "src", "index.html")})
+    ]
+
 }
