--- DmitryNikutev/src/components/Message.js
@@ -1,7 +1,0 @@
-import React from "react"
-
-export default function Message({name, text}) {
-   return (
-      <li><strong>{name}:</strong> {text}</li>
-   );
-}
