--- DmitryNikutev/src/components/MessageList.js
@@ -1,9 +1,0 @@
-import React from "react"
-import Message from "./Message";
-
-export default function MessageList({messages}) {
-   
-   return (
-      <ul>{messages.map((message, index) => <Message {...message} key={index}/>)}</ul>
-   );
-}
