--- DmitryNikutev/webpack.config.js
@@ -1,25 +1,0 @@
-const path = require("path");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-
-module.exports = {
-   entry: path.resolve(__dirname, "src", "index.js"),
-   output: {
-      path : path.resolve(__dirname, "build"),
-      filename : "index.js",
-   },
-   module: {
-      rules: [
-         {
-            test: /\.(js)$/,
-            include: path.resolve(__dirname, "src"),
-            loader: "babel-loader",
-            options: {
-               presets: ["@babel/env", "@babel/react"]
-            }
-         }
-      ]
-   },
-   plugins: [
-      new HtmlWebpackPlugin({template: path.resolve(__dirname, "src", "index.html")}),
-   ],
-};
