--- EvgenyBorodin/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "reactapp",
-  "version": "1.0.0",
-  "description": "react application",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "webpack-dev-server  --mode development --open",
-    "build": "webpack --mode production"
-  },
-  "author": "Evgeny Borodin",
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.8.6",
-    "@babel/preset-env": "^7.8.6",
-    "@babel/preset-react": "^7.8.3",
-    "babel-loader": "^8.0.6",
-    "html-webpack-plugin": "^3.2.0",
-    "webpack": "^4.42.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
-  },
-  "dependencies": {
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0"
-  }
-}
