--- EvgenyBorodin/src/message.js
@@ -1,6 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-
-export const Message = ({name, content}) => {
-    return <li><strong>{name}:</strong> {content}</li>
-}
