--- OlegShikov/src/components/Message/Message.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const Message = ({name, content}) => {
-    return <li><strong>{name}:</strong> {content}</li>
-}
