--- OlegShikov/src/components/MessageList/MessageList.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-export const MessageList = ({messages}) => {
-    return (<ul>
-        {messages.map((item, index) => <Message {...item} key={index} />)}
-    </ul>);
-}
