--- RomanOtstavnov/src/components/Message.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const Message = ({ name, content }) => (
-  <li><strong>{name}</strong>: {content}</li>
-);
-
-export default Message;
