--- RomanOtstavnov/src/index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport"
-          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <title>React</title>
-</head>
-<body>
-    <div id="app"></div>
-</body>
-</html>
