--- RomanOtstavnov/src/index.js
@@ -1,6 +1,0 @@
-import "./style.css";
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./components/App.jsx";
-
-ReactDOM.render(<App />, document.getElementById("app"));
