--- RomanOtstavnov/src/style.css
@@ -1,34 +1,0 @@
-html, body {
-  margin: 0;
-  padding: 0;
-  font-family: Arial, sans-serif;
-}
-
-.container {
-  width: 400px;
-  margin: 30px auto;
-}
-
-.message-list {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.message-list li {
-  display: block;
-  padding: 5px 0;
-  margin: 5px 0;
-  border-bottom: 1px solid #eee;
-}
-
-.message-form {
-  margin-top: 30px;
-  display: flex;
-  flex-direction: column;
-}
-
-.message-form input {
-  margin: 5px 0;
-  padding: 3px 6px;
-}
